Move currency toggle analytics from Checkout to repository layer.
